+++ libraries/javascript/utils/src/toFunction/index.ts
@@ -1,1 +1,0 @@
-export { default, type UserScript, type UserScriptParams } from './toFunction'