+++ libraries/perspective/gateway/src/main/kotlin/com/mussonindustrial/embr/perspective/gateway/component/ComponentDelegateJavaScriptProxy.kt
@@ -26,7 +26,8 @@ import org.python.core.PyObject
 
 class ComponentDelegateJavaScriptProxy(
     private val component: Component,
-    private val delegate: ComponentModelDelegate
+    private val delegate: ComponentModelDelegate,
+    private val property: String
 ) : JavaScriptProxy {
 
     private val queue = component.session.queue()
@@ -100,7 +101,7 @@ class ComponentDelegateJavaScriptProxy(
             }
             .whenCompleteAsync({ _, _ -> requestsInProgress.remove(id) }, queue::submit)
 
-        delegate.fireEvent(MESSAGE_RUN, JavaScriptRunMsg(id, function, args).getPayload())
+        delegate.fireEvent(MESSAGE_RUN, JavaScriptRunMsg(id, property, function, args).getPayload())
         return future
     }
 
@@ -160,12 +161,14 @@ class ComponentDelegateJavaScriptProxy(
 
     class JavaScriptRunMsg(
         private val id: String,
+        private val property: String,
         private val function: String,
         private val args: PyDictionary?
     ) {
         fun getPayload(): JsonObject {
             return JsonObject().apply {
                 addProperty("id", id)
+                addProperty("property", property)
                 addProperty("function", function)
                 add("args", TypeUtilities.pyToGson(args))
             }