+++ modules/charts/web/src/util/propTransforms.ts
@@ -1,37 +1,36 @@
 import {
-    isFunction,
-    toFunction,
-    readCSSVar,
-    isCSSVar,
-    PropTransform,
+  isCSSVar,
+  isFunction,
+  PropTransform,
+  readCSSVar,
+  toUserScript,
 } from '@embr-js/utils'
 
 export function getCSSTransform(
-    element: Element | null | undefined
+  element: Element | null | undefined
 ): PropTransform<unknown, unknown> {
-    return (prop: unknown) => {
-        if (
-            element !== null &&
-            element !== undefined &&
-            typeof prop === 'string' &&
-            isCSSVar(prop)
-        ) {
-            return readCSSVar(element, prop)
-        }
-        return prop
+  return (prop: unknown) => {
+    if (
+      element !== null &&
+      element !== undefined &&
+      typeof prop === 'string' &&
+      isCSSVar(prop)
+    ) {
+      return readCSSVar(element, prop)
     }
+    return prop
+  }
 }
 
 export function getScriptTransform(
-    extraContext: object = {}
+  thisArg: object = {}
 ): PropTransform<unknown, string | CallableFunction> {
-    const transform = (prop: unknown) => {
-        if (typeof prop === 'string' && isFunction(prop)) {
-            const f = toFunction(prop, extraContext)
-            return (...args: unknown[]) => f({ ...args })
-        }
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        return prop as any
+  return (prop: unknown) => {
+    if (typeof prop === 'string' && isFunction(prop)) {
+      const f = toUserScript(prop, thisArg)
+      return (...args: unknown[]) => f(...args)
     }
-    return transform
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    return prop as any
+  }
 }